Discord_bot_project-master/Discord_bot_project-master/requirements.txt
discord.py
python-dotenv
youtube_dl
